Create EnglishAuction.sol
This smart contract implements the English Auction model - the seller transfers an NFT to the auction contract, bidders place their bids and at the end the NFT is sent to the highest bidder (seller if no bidders), the top bid is send to the seller and all other bidders can withdraw their bids. 
The contract self-destructs at the end.